+++ resolwe_bio/processes/support_processors/merge_samples.py
@@ -58,8 +58,8 @@ class MergeFastqSingle(Process):
 
     slug = "merge-fastq-single"
     name = "Merge FASTQ (single-end)"
-    process_type = "data:reads:fastq:single"
-    version = "2.1.1"
+    process_type = "data:mergereads:single"
+    version = "2.2.0"
     category = "Other"
     scheduling_class = SchedulingClass.BATCH
     requirements = {
@@ -133,8 +133,8 @@ class MergeFastqPaired(Process):
 
     slug = "merge-fastq-paired"
     name = "Merge FASTQ (paired-end)"
-    process_type = "data:reads:fastq:paired"
-    version = "2.1.1"
+    process_type = "data:mergereads:paired"
+    version = "2.2.0"
     category = "Other"
     scheduling_class = SchedulingClass.BATCH
     requirements = {